+++ src/runner/desktop.rs
@@ -1,0 +1,9 @@
+use bevy::prelude::*;
+
+pub fn create_app() -> App {
+    App::new()
+}
+
+pub fn run_app(app: &mut App) -> AppExit {
+    app.run()
+}
